♻️ refactor(watchers): extract bounded-history helpers out of Docker.ts
The discovery-settle feature pushed Docker.ts over the 1600-line
structure cap, which exists to force modularization instead of
unbounded growth. Pay the toll: filterAndSliceTimestampedHistory,
appendBoundedHistoryEntry, and the two recent-history limit constants
move to a new docker-history.ts module (1602 → 1578 lines).
recent-events.test.ts exercises the helper via the module export
instead of reaching into a private method.
